Control.Monad.Skeleton

Documentation

data MonadView t m x where

A deconstructed action

Constructors

Return :: a -> MonadView t m a

(:>>=) :: !(t a) -> (a -> m b) -> MonadView t m b infixl 1

hoistMV :: (forall x. s x -> t x) -> (m a -> n a) -> MonadView s m a -> MonadView t n a

Transform the action and the continuation.

iterMV :: Monad m => (t a -> MonadView m t a) -> t a -> m a

Join MonadView recursively.

data Skeleton t a where

Skeleton t is a monadic skeleton (operational monad) made out of t. Skeletons can be fleshed out by getting transformed to other monads. It provides O(1) (>>=) and debone, the monadic reflection.

Constructors

ReturnS :: a -> Skeleton t a

BindS :: t a -> Cat (Kleisli (Skeleton t)) a b -> Skeleton t b

bone :: t a -> Skeleton t a

A skeleton that has only one bone.

debone :: Skeleton t a -> MonadView t (Skeleton t) a

Pick a bone from a Skeleton.

deboneBy :: (MonadView t (Skeleton t) a -> r) -> Skeleton t a -> r

Pick a bone from a Skeleton by a function. It's useful when used with LambdaCase.

Usecase:

 interpretM :: Monad m => Skeleton m a -> m a
 interpretM = deboneBy $ \case
   Return a -> return a
   x :>>= f -> x >>= interpretM . f

unbone :: Skeleton t a -> MonadView t (Skeleton t) a

Deprecated: Use debone instead

Uncommon synonym for debone.

boned :: MonadView t (Skeleton t) a -> Skeleton t a

Re-add a bone.

hoistSkeleton :: forall s t a. (forall x. s x -> t x) -> Skeleton s a -> Skeleton t a

Lift a transformation between bones into transformation between skeletons.